fix/#2u9czt: refactor name/location for a class
