+++ src/main/java/proj/kedabra/billsnap/presentation/resources/BillResource.java
@@ -11,7 +11,6 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import proj.kedabra.billsnap.business.utils.enums.BillStatusEnum;
-import proj.kedabra.billsnap.utils.tuples.AccountStatusCompletePair;
 
 @Data
 public class BillResource implements Serializable {
@@ -51,7 +50,7 @@ public class BillResource implements Serializable {
     private List<ItemResource> items;
 
     @ApiModelProperty(name = "List of accounts associated to the bill", position = 10)
-    private List<AccountStatusCompletePair> accountsList;
+    private List<AccountStatusResource> accountsList;
 
     @ApiModelProperty(name = "the total amount of the bill", position = 11)
     private BigDecimal balance;
